added version rolling frequency setting

+++ components/bm1397/include/asic.h
@@ -106,11 +106,11 @@ public:
     virtual uint8_t nonceToAsicNr(uint32_t nonce) = 0;
 
     void setVrFreqReg(uint32_t value);
-    void setVrFrequency(float freq);
+    void setVrFrequency(uint32_t freq);
 
     // static helper functions
-    static uint32_t vrFreqToReg(float freq_hz);
-    static float vrRegToFreq(uint32_t reg);
+    static uint32_t vrFreqToReg(uint32_t freq_hz);
+    static uint32_t vrRegToFreq(uint32_t reg);
 
     // asic models specific
     virtual uint8_t init(uint64_t frequency, uint16_t asic_count, uint32_t difficulty, uint32_t vrFreqReg) = 0;
